backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
ta==0.10.2
python-binance==1.0.19